pyproject.toml
[tool.poetry]
name = "arcanum-app"
version = "0.1.0"
description = ""
authors = ["Arthur Ratsa <rheni.light@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
python-dotenv = "^1.1.0"
python-dateutil = "^2.9.0.post0"
pytz = "^2025.2"
colorlog = "^6.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
